sound: bundle sounds in resources
With WebKitGTK on Linux the sounds need to play from a GStreamer
backend. This requires access to the files from an absolute filesystem
path, as provided through the `asset://` protocol. Bundle the sounds in
the resources instead of frontendDist, and retrieve the path through
`resolveResource`.

This also reuses a single Audio element to play the sounds, with major
benefits for the application behavior when spamming sound effects by
e.g. scrolling through board positions. On WebKitGTK, the old behavior
easily causes the application to crash with the limit on open file
descriptors reached. Other web engines may behave better already.

Signed-off-by: Stefan Nuernberger <kabelfrickler@gmail.com>